+++ R/generate_questions.R
@@ -2,6 +2,7 @@
 #'
 #' @param x List to get the params for UI generation
 #' @param inputId Character string with inputId if not in the x
+#' @param module_ui_id Character string getting the id for module UI
 #'
 #' @import shiny
 #' @importFrom stats setNames
@@ -76,9 +77,11 @@
 #'
 #' @param source_list list object with inputs parameters
 #' @param div_id Character string declaring id for the div
-#' @param module_ui_id character string declaring module id
 #' @param css Character string containing custom css rules for classes
 #' 'mandatory_star' and 'invalid_input'
+#' @param button_label character string with active lavel for the submission
+#' button
+#' @param module_ui_id character string declaring module id
 #' @import shiny
 #' @import htmltools