+++ R/survey_module.R
@@ -1,5 +1,7 @@
 #' Create UI of your questionnaire
 #'
+#' @param module_id Character string holding ID for the module. Needs to be the
+#' same as one provided for 'quetzio_server'
 #' @import shiny
 #' @export
 #'
@@ -206,11 +208,12 @@ quetzio_server <- R6::R6Class(
 
       # save the module id into environment
 
-      self$module_id <- .null_def(module_id, uuid::UUIDgenerate())
-      if (is.null(link_id))  {
+      self$module_id <- module_id
+
+      if (is.null(add_args$link_id))  {
         self$module_ui_id <- self$module_id
       } else {
-        self$module_ui_id <- paste(link_id, self$module_id, sep = ns.sep)
+        self$module_ui_id <- paste(add_args$link_id, self$module_id, sep = ns.sep)
       }
       # read the file and save resulting list in the environment